+++ packages/Bio/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Bio changelog
 
+## 2.11.38 (2024-03-08)
+
+### Bug fixes
+
+* GROK-14910: PepSeA verbose output
+* MSA ensures docker container for PepSeA
+* Sample files harmonized
+
+## 2.11.37 (2024-03-08)
+
+### Bug fixes
+
+* Check Bio publishing, PepSeA docker
+
 ## 2.11.36 (2024-02-28)
 
 ### Bug fixes